+++ R/completion_status.R
@@ -18,86 +18,89 @@
 #'
 #' @param dframe `r dframe` with required variables `{mcid, timely_term}.`
 #'
-#' @param midfield_table `r midfield_x("degree")` with required
+#' @param midf_table `r midfield_x("degree")` with required
 #'        variables `{mcid, term_degree}.`
 #'
 #' @returns Data frame with the following properties:
 #' * `r df_class_preserved`
 #' * `r rows_not_modified`
 #' * `r new_cols_added`
-#'   - `term_degree` &nbsp; Joined from `midfield_table.`
+#'   - `term_degree` &nbsp; Joined from `midf_table.`
 #'   - `completion_status` &nbsp; Character. Possible values of "timely",
 #'      "late" and "NA".
 #'
 #' @example man/examples/exa_completion_status.R
 #' @export
 #'
-completion_status <- function(dframe, midfield_table = degree) {
+completion_status <- function(dframe, midf_table = degree) {
   #
   # ---------- assign active column names
-
+  
   reqd_dframe_vars <- c("mcid", "timely_term")
   reqd_table_vars <- c("mcid", "term_degree")
   added_vars <- c("term_degree", "completion_status")
-
+  
   # ---------- base R checks (all data frame classes)
-
+  
   # data frame assessment
   qassert(dframe, "d+")
-  qassert(midfield_table, "d+")
-
+  qassert(midf_table, "d+")
+  
   # required columns
   assert_names(colnames(dframe), must.include = reqd_dframe_vars)
-  assert_names(colnames(midfield_table), must.include = reqd_table_vars)
-
+  assert_names(colnames(midf_table), must.include = reqd_table_vars)
+  
   # class of required columns
-  for (i in seq_along(reqd_dframe_vars)) {
-    qassert(dframe[[reqd_dframe_vars[i]]], "s+")
-  }
-  for (i in seq_along(reqd_table_vars)) {
-    qassert(midfield_table[[reqd_table_vars[i]]], "s+")
-  }
-
+  for (var in reqd_dframe_vars) qassert(dframe[[var]], "s+")
+  for (var in reqd_table_vars) qassert(midf_table[[var]], "s+")
+  
   # ---------- preparation
-
-  # to restore class, but not grouped_DF (tibbles)
+  
+  # to restore class except grouped tibbles
   prior_class <- setdiff(class(dframe), "grouped_df")
-
+  
   # prevent by-ref changes propagating to global env
   dframe <- copy(dframe)
   setDT(dframe)
-  midf_table <- copy(midfield_table)
+  midf_table <- copy(midf_table)
   setDT(midf_table)
-
+  
+  # avoid overwriting columns that match names of temporary columns
+  init_temp_vars <- c("idx")
+  temp_vars <- edit_new_col_names(dframe, init_temp_vars)
+  idx_chr <- temp_vars[1]
+  
   # bind names due to NSE notes in R CMD check
   completion_status <- NULL
-  idx <- NULL
   timely_term <- NULL
-
+  IDX <- NULL
+  
   # ---------- do the work
-
+  
   # save columns except those being added
   saved_vars <- setdiff(colnames(dframe), added_vars)
   return_vars <- c(saved_vars, added_vars)
-
+  
   # drop added vars, omit NA in required vars
   dframe <- dframe[, .SD, .SDcols = saved_vars]
   dframe <- na.omit(dframe, cols = reqd_dframe_vars)
   dframe <- unique(dframe)
-
+  
   # keep required vars and omit NAs
   midf_table <- midf_table[, .SD, .SDcols = reqd_table_vars]
   midf_table <- na.omit(midf_table, cols = reqd_table_vars)
   midf_table <- unique(midf_table)
-
+  
   # add temp col to restore row order
-  dframe[, idx := .I]
-
+  dframe[, IDX := .I,
+         env = list(IDX = idx_chr)
+  ]
+  
   # join degree records
   dframe <- midf_table[dframe, on = "mcid"]
-
+  
   # ---------- timely completion labels
-
+  
   # completion is timely, late, or NA
   dframe[, completion_status := fifelse(
     term_degree <= timely_term,
@@ -105,21 +108,21 @@ completion_status <- function(dframe, midfield_table = degree) {
     "late",
     na = NA_character_
   )]
-
+  
   # ---------- prepare to return
-
+  
   # restore row order
-  setkey(dframe, idx)
-
+  setkeyv(dframe, idx_chr)
+  
   # drop temporary cols, restore original col order
   dframe <- dframe[, .SD, .SDcols = return_vars]
-
+  
   # ensure unique rows
   dframe <- unique(dframe)
-
+  
   # restore class
   setattr(dframe, "class", prior_class)
-
+  
   # done
   dframe[]
 }
@@ -139,7 +142,7 @@ add_completion_status <- function(dframe, midfield_degree = degree) {
     msg = "This function was deprecated as part of an update to all
     midfieldr functions. Please use `completion_status()` instead."
   )
-
+  
   # original function calls the new function
-  completion_status(dframe = dframe, midfield_table = midfield_degree)
+  completion_status(dframe = dframe, midf_table = midfield_degree)
 }