+++ R/filter_programs.R
@@ -1,41 +1,5 @@
 # See R/roxygen.R for documentation below that uses inline R code
-#
-# ---------- deprecated version
-#
-#' midfieldr deprecated functions
-#' @param keep_text Deprecated `filter_cip()`. Character vector of search text to keep.
-#' @param drop_text Deprecated `filter_cip()`. Character vector of search text to drop.
-#' @param cip Deprecated `filter_cip()`. Data frame of programs to be searched.
-#' @param select Deprecated `filter_cip()`. Character vector of column names to select.
-#' @rdname midfieldr-deprecated
-#' @export
-filter_cip <- function(keep_text = NULL,
-                       drop_text = NULL,
-                       cip = NULL,
-                       select = NULL) {
-  .Deprecated(
-    new = "filter_programs",
-    package = "midfieldr",
-    msg = "This function was deprecated to put the data frame as the
-    first argument making it possible to chain with other functions.
-    Please use `filter_programs()` instead."
-  )
 
-  # attempt to continue to use original function with partial success
-  if (is.null(cip)) cip <- midfieldr::cip
-  negate <- FALSE
-
-  if (!is.null(drop_text)) {
-    keep_text <- drop_text
-    negate <- TRUE
-  }
-
-  filter_programs(dframe = cip, pattern = keep_text, negate = negate)
-}
-NULL
-#
-# ---------- current version
-#
 #' Choose rows of CIP data
 #'
 #' Subset a CIP data frame, retaining rows that match or partially match
@@ -82,7 +46,7 @@ filter_programs <- function(dframe, pattern, ..., negate = FALSE) {
 
   # ---------- preparation
 
-  # to restore class, but not grouped_DF (tibbles)
+  # to restore class except grouped tibbles
   prior_class <- setdiff(class(dframe), "grouped_df")
 
   # prevent by-ref changes propagating to global env
@@ -109,9 +73,47 @@ filter_programs <- function(dframe, pattern, ..., negate = FALSE) {
   dframe <- unique(dframe)
 
   # restore class
-  setkey(dframe, NULL)
   setattr(dframe, "class", prior_class)
 
   # done
   dframe[]
+}
+
+
+
+
+
+
+
+# ---------- deprecated version
+#
+#' midfieldr deprecated functions
+#' @param keep_text Deprecated `filter_cip()`. Character vector of search text to keep.
+#' @param drop_text Deprecated `filter_cip()`. Character vector of search text to drop.
+#' @param cip Deprecated `filter_cip()`. Data frame of programs to be searched.
+#' @param select Deprecated `filter_cip()`. Character vector of column names to select.
+#' @rdname midfieldr-deprecated
+#' @export
+filter_cip <- function(keep_text = NULL,
+                       drop_text = NULL,
+                       cip = NULL,
+                       select = NULL) {
+  .Deprecated(
+    new = "filter_programs",
+    package = "midfieldr",
+    msg = "This function was deprecated to put the data frame as the
+    first argument making it possible to chain with other functions.
+    Please use `filter_programs()` instead."
+  )
+
+  # attempt to continue to use original function with partial success
+  if (is.null(cip)) cip <- midfieldr::cip
+  negate <- FALSE
+
+  if (!is.null(drop_text)) {
+    keep_text <- drop_text
+    negate <- TRUE
+  }
+
+  filter_programs(dframe = cip, pattern = keep_text, negate = negate)
 }