+++ R/timely_term.R
@@ -20,7 +20,7 @@
 #'
 #' @param dframe `r dframe` with required variable `{mcid}.`
 #'
-#' @param midfield_table `r midfield_x("term")` with required variables
+#' @param midf_table `r midfield_x("term")` with required variables
 #'        `{mcid, term, level}.`
 #'
 #' @param ... `r param_dots`
@@ -37,9 +37,9 @@
 #' * `r rows_not_modified`
 #' * `r new_cols_added`
 #'   - `term_i` &nbsp; Character. Initial term of a student's longitudinal
-#'      record, encoded `YYYYT`. Extracted from `midfield_table.`
+#'      record, encoded `YYYYT`. Extracted from `midf_table.`
 #'   - `level_i` &nbsp; Character. Student level (01 Freshman, 02 Sophomore,
-#'      etc.) in their initial term. Extracted from `midfield_table.`
+#'      etc.) in their initial term. Extracted from `midf_table.`
 #'   - `adj_span` &nbsp; Numeric. Integer span of years for timely
 #'      completion adjusted for a student's initial level.
 #'   - `timely_term` &nbsp; Character. Latest term by which program completion
@@ -49,7 +49,7 @@
 #' @export
 #'
 timely_term <- function(dframe,
-                        midfield_table = term,
+                        midf_table = term,
                         ...,
                         sched_span = NULL,
                         span = NULL) {
@@ -70,19 +70,15 @@ timely_term <- function(dframe,
 
   # data frame assessment
   qassert(dframe, "d+")
-  qassert(midfield_table, "d+")
+  qassert(midf_table, "d+")
 
   # required columns
   assert_names(colnames(dframe), must.include = reqd_dframe_vars)
-  assert_names(colnames(midfield_table), must.include = reqd_table_vars)
+  assert_names(colnames(midf_table), must.include = reqd_table_vars)
 
   # class of required columns
-  for (i in seq_along(reqd_dframe_vars)) {
-    qassert(dframe[[reqd_dframe_vars[i]]], "s+")
-  }
-  for (i in seq_along(reqd_table_vars)) {
-    qassert(midfield_table[[reqd_table_vars[i]]], "s+")
-  }
+  for (var in reqd_dframe_vars) qassert(dframe[[var]], "s+")
+  for (var in reqd_table_vars) qassert(midf_table[[var]], "s+")
 
   # other arguments
   span <- span %?% 6
@@ -93,23 +89,32 @@ timely_term <- function(dframe,
 
   # ---------- preparation
 
-  # to restore class except for groups in tibbles
+  # to restore class except grouped tibbles
   prior_class <- setdiff(class(dframe), "grouped_df")
 
   # prevent by-ref changes propagating to global env
   dframe <- copy(dframe)
   setDT(dframe)
-  midf_table <- copy(midfield_table)
+  midf_table <- copy(midf_table)
   setDT(midf_table)
+
+  # avoid overwriting columns that match names of temporary columns
+  init_temp_vars <- c("idx", "yyyy", "t", "delta")
+  temp_vars <- edit_new_col_names(dframe, init_temp_vars)
+  idx_chr   <- temp_vars[1]
+  yyyy_chr  <- temp_vars[2]
+  term_chr  <- temp_vars[3]
+  delta_chr <- temp_vars[4]
 
   # bind names due to NSE notes in R CMD check
   adj_span <- NULL
-  delta <- NULL
-  idx <- NULL
   level_i <- NULL
   term_i <- NULL
   timely_term <- NULL
-  yyyy <- NULL
+  IDX <- NULL
+  DELTA <- NULL
+  TERM <- NULL
+  YYYY <- NULL
 
   # ---------- do the work
 
@@ -128,7 +133,7 @@ timely_term <- function(dframe,
   midf_table <- unique(midf_table)
 
   # add temp col to restore row order
-  dframe[, idx := .I]
+  dframe[, IDX := .I, env = list(IDX = idx_chr)]
 
   # inner-join dframe ID and required table vars
   x <- unique(dframe[, .(mcid)])
@@ -137,8 +142,8 @@ timely_term <- function(dframe,
 
   # keep the row of the first term, lowest level, by ID and institution
   setorderv(x, c("mcid", "term"))
-  x <- x[, .SD[1], by = c("mcid")]
-
+  x <- x[, .SD[1L], by = c("mcid")]
+  
   # rename term and level
   x <- x[, .(mcid, term_i = term, level_i = level)]
 
@@ -148,36 +153,65 @@ timely_term <- function(dframe,
   # ---------- construct timely term
 
   dframe[, `:=`(
-    yyyy = substr(term_i, 1, 4),
-    t    = substr(term_i, 5, 5)
-  )]
+    YYYY = substr(term_i, 1, 4),
+    TERM = substr(term_i, 5, 5)
+  ),
+  env = list(
+    YYYY = yyyy_chr,
+    TERM = term_chr
+  )
+  ]
 
-  # for month terms, (letters A, B, C, ...), set first term to zero
-  dframe <- dframe[t %chin% LETTERS | t %chin% letters, t := "0"]
+  # for month terms, (letters A, B, ..., a, b, ...), set first term to zero
+  dframe <- dframe[TERM %chin% c(LETTERS, letters), TERM := "0",
+                   env = list(TERM = term_chr)
+  ]
 
   # make year and term numeric
-  dframe[, names(.SD) := lapply(.SD, as.numeric), .SDcols = c("yyyy", "t")]
+  dframe[, names(.SD) := lapply(.SD, as.numeric), .SDcols = c(yyyy_chr, term_chr)]
 
   # if first term is in summer, delay to the subsequent Fall
-  dframe[t > 3, `:=`(yyyy = yyyy + 1, t = 1)]
+  dframe[TERM > 3, `:=`(
+    YYYY = YYYY + 1,
+    TERM = 1
+  ),
+  env = list(
+    YYYY = yyyy_chr,
+    TERM = term_chr
+  )
+  ]
 
   # reduce span by assumed number of completed years by level
-  dframe[, delta := fcase(
-    level_i %like% "04", 3,
+  dframe[, DELTA := fcase(level_i %like% "04", 3,
     level_i %like% "03", 2,
     level_i %like% "02", 1,
     default = 0
-  )]
-  dframe[, adj_span := span - delta]
+  ),
+  env = list(DELTA = delta_chr)
+  ]
+
+  dframe[, adj_span := span - DELTA,
+    env = list(DELTA = delta_chr)
+  ]
 
   # use adj_span to construct estimated timely-completion term
-  dframe[t == 0 | t == 1, timely_term := paste0(yyyy + adj_span - 1, 3)]
-  dframe[t > 1, timely_term := paste0(yyyy + adj_span, 1)]
+  dframe[TERM == 0 | TERM == 1, timely_term := paste0(YYYY + adj_span - 1, 3),
+         env = list(
+           YYYY = yyyy_chr,
+           TERM = term_chr
+         )
+  ]
+  dframe[TERM > 1, timely_term := paste0(YYYY + adj_span, 1),
+    env = list(
+      YYYY = yyyy_chr,
+      TERM = term_chr
+    )
+  ]
 
   # ---------- prepare to return
 
   # restore row order
-  setkey(dframe, idx)
+  setkeyv(dframe, idx_chr)
 
   # drop temporary cols, restore original col order
   dframe <- dframe[, .SD, .SDcols = return_vars]
@@ -217,7 +251,7 @@ add_timely_term <- function(dframe,
   # original function calls the new function
   timely_term(
     dframe = dframe,
-    midfield_table = midfield_term,
+    midf_table = midfield_term,
     ...,
     sched_span = sched_span,
     span = span