+++ R/utils.R
@@ -89,156 +89,26 @@ sort_uniq <- function(x,
 
 # ------------------------------------------ INTERNAL UTILITIES
 
-#' Set column order and row order
+#' Edit names of new columns
 #'
-#' Use the vector of column names in `cols` as the ordering argument in
-#' `data.table::setcolorder()`` and as the key argument in
-#' `data.table::setkeyv()` to order the rows.
-#'
-#' @param dframe data frame
-#' @param cols character vector of column names to use as keys
-#' @noRd
-#'
-set_colrow_order <- function(dframe, cols) {
-  on.exit(setkey(dframe, NULL))
-
-  # ensure dframe is data.table class
-  setDT(dframe)
-
-  # column order of data frame by vector of names
-  setcolorder(dframe, neworder = cols)
-
-  # order rows by using names as keys
-  setkeyv(dframe, cols = cols)
-}
-
-# ------------------------------------------------------------------------
-#
-#' Return column names not overwritten by the function
-#'
-#' Identifies the names of columns unaffected by the function operation.
-#' Used by several midfieldr "add_" functions.
-#'
-#' @param dframe data frame
-#' @param new_cols character vector of column names added by the function
-#' @noRd
-#'
-find_old_cols <- function(dframe, new_cols) {
-  all_cols <- colnames(dframe)
-  old_cols <- all_cols[!all_cols %chin% new_cols]
-  return(old_cols)
-}
-
-
-# ------------------------------------------------------------------------
-#
-#' Add a column of institution names
-#'
-#' Add a column of character values with institution names (or labels) using
-#' student ID as the join-by variable. Obtains the information from the MIDFIELD
-#' `term` data table or equivalent. In the MIDFIELD practice data, the labels
-#' are de-identified.
-#'
-#' If a student is associated with more than one institution, the institution at
-#' which they completed the most terms is returned. An existing column with the
-#' same name as the added column is overwritten.
-#'
-#' @param dframe Data frame with required variable `mcid.`
-#' @param midfield_table MIDFIELD `term` data table or equivalent with required
-#' variables `mcid`, `institution`, and `term`.
-#' @noRd
-#'
-add_institution <- function(dframe,
-                            midfield_table = term) {
-  # remove all keys
-  on.exit(setkey(dframe, NULL))
-  on.exit(setkey(midfield_table, NULL), add = TRUE)
-
-  # required arguments
-  qassert(dframe, "d+")
-  qassert(midfield_table, "d+")
-
-  # optional arguments
-  # NA
-
-  # inputs modified (or not) by reference
-  dframe <- copy(as.data.table(dframe)) #  must copy
-  setDT(midfield_table) # immediately subset, so side-effect OK
-
-  # required columns
-  assert_names(colnames(dframe),
-    must.include = c("mcid")
-  )
-  assert_names(colnames(midfield_table),
-    must.include = c("mcid", "institution", "term")
-  )
-
-  # class of required columns
-  qassert(dframe[, mcid], "s+")
-  qassert(midfield_table[, mcid], "s+")
-  qassert(midfield_table[, institution], "s+")
-  qassert(midfield_table[, term], "s+")
-
-  # bind names due to NSE notes in R CMD check
-  N <- NULL
-
-  # do the work
-  # Inner join using three columns of term
-  x <- midfield_table[, .(mcid, institution, term)]
-  y <- unique(dframe[, .(mcid)])
-  DT <- y[x, on = .(mcid), nomatch = NULL]
-
-  # count terms at institutions
-  DT <- DT[, .N, by = c("mcid", "institution")]
-
-  # what if there is a tie? can we select the most recent institution?
-  # keep the institution with the most terms (if more than one)
-  setkeyv(DT, c("mcid", "N"))
-  DT <- DT[, .SD[.N], by = "mcid"]
-  DT[, N := NULL]
-
-  # join to dframe, overwrite institution if any
-  if ("institution" %chin% names(dframe)) {
-    dframe[, institution := NULL]
-  }
-
-  # left join, keep all rows of dframe
-  setkeyv(DT, "mcid")
-  setkeyv(dframe, "mcid")
-  dframe <- DT[dframe]
-  return(dframe)
-}
-
-
-# ------------------------------------------------------------------------
-#
-#' Setup vector of column names for data frame return
-#'
-#' Several midfieldr functions add new columns to an existing data frame.
-#' Sometimes one or more of the new columns are already extant in the data
-#' frame. This function sorts the column names such that new-but-extant
-#' columns are overwritten but do not change position and all other new
-#' columns are added to the right of the data frame.
+#' Prevents overwriting existing columns in data frame that happen to match
+#' internal, temporary column names within the function.
 #'
 #' @param dframe Data frame to which columns are being added.
-#' @param active_cols vector of column names being added/overwritten
-#'        by the function.
+#' @param proposed_new_names Vector of column names being added internally
+#'        in the function, excluding the names of columns that are the output
+#'        of the function.
 #' @noRd
-setup_return_cols <- function(dframe, active_cols) {
-  # columns of incoming dframe
-  orig_cols <- colnames(dframe)
+edit_new_col_names <- function(dframe, proposed_new_names) {
+  # prevent any possible by-ref changes (probably not necessary here)
+  dframe <- copy(dframe)
 
-  # new columns present in the original dframe
-  active_orig <- intersect(active_cols, orig_cols)
+  # existing column names
+  exist_names <- colnames(dframe)
 
-  # new columns not present in the original dframe
-  new_cols_non_orig <- setdiff(active_cols, active_orig)
+  # add suffix .1, .2, etc. to new names as needed if they match existing
+  uniq_names <- make.unique(c(exist_names, proposed_new_names))
 
-  # columns to return in order
-  return_cols <- c(orig_cols, new_cols_non_orig)
-
-  # rearrange so that active cols always added to the right
-  inactive_cols <- setdiff(return_cols, active_cols)
-
-  return_cols <- c(inactive_cols, active_cols)
+  # return the edited new names
+  new_names <- setdiff(uniq_names, exist_names)[]
 }