+++ man/examples/exa_post_completion_terms.R
@@ -3,12 +3,12 @@ term <- select_basic_cols(toy_term)
 degree <- select_basic_cols(toy_degree)
 
 # identify term-clusters in a 'term' table
-x <- post_bacc_terms(term, degree)
+x <- post_completion_terms(term, degree)
 x[, .N, by = "term_cluster"][order(-N)]
 x
 
 # identify term-clusters in a 'degree' table
-x <- post_bacc_terms(degree, degree)
+x <- post_completion_terms(degree, degree)
 x[, .N, by = "term_cluster"][order(-N)]
 x