+++ man/examples/exa_timely_term.R
@@ -5,9 +5,9 @@ x <- toy_student[c(51:55, 346:350), .(mcid, sex)]
 x
 
 # Add timely term columns. Unrelated columns (sex) are unaffected.
-x <- timely_term(x, midfield_table = term)
+x <- timely_term(x, midf_table = term)
 x
 
 # Repeat. New columns silently replace existing columns of the same name.
-y <- timely_term(x, midfield_table = term)
+y <- timely_term(x, midf_table = term)
 y