+++ NEWS.md
@@ -18,7 +18,7 @@
 
 New features
 
-- New function `post_bacc_terms()` to identify rows of post-baccalaureate terms to exclude.
+- New function `post_completion_terms()` to identify rows of post-baccalaureate terms to exclude.
 - New utilities  `look_at()`, `catch_error()`, and `sort_uniq()` that wrap base R functions with our preferred arguments. 
 - Add data set `cip2010`, constructed from a recent download of the source file (`CIPCode2010.csv`) from NCES. Has the same data structure as the existing `cip` dataset, but with more rows. 
 - Renamed/revised functions and their arguments:
@@ -39,6 +39,7 @@ Bug fixes
 
 - In vignettes, fixed `.SD[]` where finding the first instance of a term should return more than one row, e.g., graduating with a double degree. 
 - This same fix added two rows to the `study_observations` built-in data set. 
+- Existing column names that happen to have the same name as internal temporary column names are now prevented from being overwritten.